modified:   jobs/JGLOBAL_NCEPPOST, scripts/exgdas_nceppost.sh.ecf, scripts/exgfs_nceppost.sh.ecf,  ush/gfs_nceppost.sh
Add an option INLINE_POST.  If it is true, the grib2 "master" file is produced by the inline post in the
forecast model.  gfs_nceppost.sh script is then skipped.  Also set FILTER=0 to turn off 5wav-filtered data.
The UPP still needs to be modified to read in netcdf history files to produce sflux, goes image and gtg products.